third_party/blink/web_tests/wpt_internal/document-transition/parsing/page-transition-tag-invalid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Document Transitions: page-transition-tag with invalid values</title>
<link rel="author" title="Vladimir Levin" href="mailto:vmpstr@chromium.org">
<link rel="help" href="https://github.com/WICG/shared-element-transitions">
<meta name="assert" content="page-transition-tag does not support reserved tag names.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
// 'root' is a reserved tag.
test_invalid_value("page-transition-tag", "root");
</script>
</body>
</html>
